refactor : [ #5 모듈구조 변경 ]
