+++ JwtService/src/main/resources/application-local.yml
@@ -1,6 +1,17 @@
+#spring:
+#  datasource:
+#    url: jdbc:mariadb://localhost:3306/jwt_db
+#    username: root
+#    password: root1234
+#    driver-class-name: org.mariadb.jdbc.Driver
+
 spring:
   datasource:
-    url: jdbc:mariadb://localhost:3306/jwt_db
-    username: root
-    password: root
-    driver-class-name: org.mariadb.jdbc.Driver
+    url: jdbc:h2:tcp://localhost/~/member_db
+    username: sa
+    password:
+    driver-class-name: org.h2.Driver
+
+  output:
+    ansi:
+      enabled: always
