+++ JwtService/src/main/resources/application.yml
@@ -15,7 +15,7 @@ spring:
       hibernate:
         format_sql: true
         show_sql: true
-        dialect: org.hibernate.dialect.MariaDBDialect
+#        dialect: org.hibernate.dialect.MariaDBDialect
   application:
     name: jwt-service
   batch:
