+++ MemberService/src/main/java/com/memberservice/member/model/dto/jackson/Views.java
@@ -1,4 +1,4 @@
-package com.memberservice.member.dto.jackson;
+package com.memberservice.member.model.dto.jackson;
 
 public class Views {
     public static class Basic {}
