+++ MemberService/src/main/java/com/memberservice/member/model/entity/BaseEntity.java
@@ -1,4 +1,4 @@
-package com.memberservice.member.entity;
+package com.memberservice.member.model.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.EntityListeners;
