+++ MemberService/src/main/java/com/memberservice/relationship/repository/RelationshipRepository.java
@@ -1,0 +1,11 @@
+//package com.memberservice.relationship.repository;
+//
+//import com.memberservice.relationship.entity.Relationship;
+//import com.memberservice.relationship.entity.RelationshipPK;
+//import org.springframework.data.jpa.repository.JpaRepository;
+//import org.springframework.stereotype.Repository;
+//
+//@Repository
+//public interface RelationshipRepository extends JpaRepository<Relationship, RelationshipPK> {
+//
+//}
