+++ MemberService/src/main/resources/application-local.yml
@@ -1,9 +1,13 @@
-#spring:
-#  datasource:
-#    url: jdbc:h2:tcp://localhost/~/member_db
-#    username: sa
-#    password:
-#    driver-class-name: org.h2.Driver
+spring:
+  datasource:
+    url: jdbc:h2:tcp://localhost/~/member_db
+    username: sa
+    password:
+    driver-class-name: org.h2.Driver
+
+  ansi:
+    output:
+      enabled: always
 
 #spring:
 #  datasource:
@@ -12,9 +16,9 @@
 #    password: root1234
 #    driver-class-name: org.mariadb.jdbc.Driver
 
-spring:
-  datasource:
-    url: jdbc:mariadb://localhost:3306/member_db
-    username: root
-    password: root
-    driver-class-name: org.mariadb.jdbc.Driver
+#spring:
+#  datasource:
+#    url: jdbc:mariadb://localhost:3306/testdb
+#    username: root
+#    password: root
+#    driver-class-name: org.mariadb.jdbc.Driver
