+++ MemberService/src/main/resources/application.yml
@@ -10,10 +10,15 @@ spring:
   autoconfigure:
     exclude: org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration
 
+#  h2:
+#    console:
+#      enabled: false
+#      settings:
+#        web-allow-others: false
 
   jpa:
     hibernate:
-      ddl-auto: create-drop
+      ddl-auto: create
     properties:
       hibernate:
         format_sql: true
@@ -21,11 +26,9 @@ spring:
 #        dialect: org.hibernate.dialect.MariaDBDialect
 
   main:
-    allow-bean-definition-overriding: false
+    allow-bean-definition-overriding: true
 
 # TRACE(상세) > DEBUG(디버깅) > INFO(정보) > WARN(경고) > ERROR(에러) > FATAL(치명적 에러) > OFF(끄기)
-logging:
-  level:
-    org:
-      springframework:
-        security: WARN
+#logging:
+#  level:
+#    root: off
